Logging has been adjusted and improved
