+++ plume/src/test/resources/log4j2.properties
@@ -5,24 +5,26 @@ property.filename=${sys:java.io.tmpdir}/plume/plume-test.log
 filters=threshold
 filter.threshold.type=ThresholdFilter
 filter.threshold.level=debug
-appenders=rolling
-appender.rolling.type=RollingFile
-appender.rolling.name=RollingFile
-appender.rolling.fileName=${filename}
-appender.rolling.filePattern=${sys:java.io.tmpdir}/plume/debug-backup-%d{MM-dd-yy-HH-mm-ss}-%i.log.gz
-appender.rolling.layout.type=PatternLayout
-appender.rolling.layout.pattern=%d{HH:mm:ss} %-5p %c{1}:%L - %m%n
-appender.rolling.policies.type=Policies
-appender.rolling.policies.time.type=TimeBasedTriggeringPolicy
-appender.rolling.policies.time.interval=1
-appender.rolling.policies.time.modulate=true
-appender.rolling.policies.size.type=SizeBasedTriggeringPolicy
-appender.rolling.policies.size.size=10MB
-appender.rolling.strategy.type=DefaultRolloverStrategy
-appender.rolling.strategy.max=20
-loggers=rolling
+appenders=plume
+appender.plume.type=RollingFile
+appender.plume.name=RollingFile
+appender.plume.fileName=${filename}
+appender.plume.filePattern=${sys:java.io.tmpdir}/plume/debug-backup-%d{MM-dd-yy-HH-mm-ss}-%i.log.gz
+appender.plume.layout.type=PatternLayout
+appender.plume.layout.pattern=%d{HH:mm:ss} %-5p %c{1}:%L - %m%n
+appender.plume.policies.type=Policies
+appender.plume.policies.time.type=TimeBasedTriggeringPolicy
+appender.plume.policies.time.interval=1
+appender.plume.policies.time.modulate=true
+appender.plume.policies.size.type=SizeBasedTriggeringPolicy
+appender.plume.policies.size.size=10MB
+appender.plume.strategy.type=DefaultRolloverStrategy
+appender.plume.strategy.max=20
+loggers=plume,tigergraph
 #Make sure to change the package structure as per your application
-logger.rolling.name=io.github.plume
-logger.rolling.level=debug
-logger.rolling.additivity=false
-logger.rolling.appenderRef.rolling.ref=RollingFile
+logger.plume.name=io.github.plume
+logger.plume.level=debug
+logger.plume.additivity=false
+logger.plume.appenderRef.rolling.ref=RollingFile
+logger.tigergraph.name=com.tigergraph
+logger.tigergraph.level=fatal
